Add checks
The avatar command seemed to not work when there was no user specified, the new code in lines 11 - 13 guarantee there will be a user that the avatar command will work on. The "click to see avatar" link will also try to now get the avatar at a high resolution with a query param, the main image of the embed will be 1024x1024 so it doesn't take up the whole chat.