ci(CodeQL): actually ignore all events from dependabot (#290)
Instead of ignoring all pushes to dependabot branches, this commit changes the CodeQL workflow's behavior so that all events from dependabot are ignored, in order to completely remove the need to re-run failed workflows. Additionally, pushes to dependabot branches by people who aren't dependabot are not going to be ignored anymore.